linting sweep. reorganized tests.
