--- src/__tests__/background.test.js
@@ -1,7 +1,0 @@
-
-// test that is irrelevent
-describe('title', () => {
-    it('has an h1 tag with the text cypress recorder', () => {
-      expect(true).toBe(true);
-    })
-})
