--- src/__tests__/codeGenerator.test.js
@@ -1,7 +1,0 @@
-// unit tests for code generator
-// test that is irrelevent
-describe('title', () => {
-  it('has an h1 tag with the text cypress recorder', () => {
-    expect(true).toBe(true);
-  })
-})
