--- src/__tests__/eventRecorder.test.js
@@ -1,15 +1,0 @@
-// unit tests for e recorder
-// const recorder = require("../event-recording/EventRecorder.ts");
-
-// const sum = require("../event-recording/EventRecorder.ts")
-
-// test("adds that shit up", () => {
-//     expect(sum(1,2)).toBe(3);
-// });
-
-// test that is irrelevent
-describe('title', () => {
-    it('has an h1 tag with the text cypress recorder', () => {
-        expect(true).toBe(true);
-    })
-})
