+++ src/__tests__/toggleButton.test.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import { shallow } from 'enzyme';
-import ToggleButton from '../../popup/components/ToggleButton';
-import '../../setupTests';
-import { ControlAction } from '../../constants';
+import ToggleButton from '../popup/components/ToggleButton';
+import { ControlAction } from '../constants';
+import '../setupTests';
 
 describe('toggle button', () => {
   let wrapper;
@@ -11,31 +11,31 @@ describe('toggle button', () => {
     handleToggleMock = jest.fn();
   });
   it('should not call handletoggle when clicked and recStatus is "off" and tab is invalid', () => {
-    wrapper = shallow(<ToggleButton recStatus='off' isValidTab={false} handleToggle={handleToggleMock}/>)
+    wrapper = shallow(<ToggleButton recStatus="off" isValidTab={false} handleToggle={handleToggleMock} />);
     expect(wrapper).toMatchSnapshot();
     const button = wrapper.find('button');
     expect(button.text()).toBe('Invalid Tab');
     expect(button.prop('disabled')).toBe(true);
   });
   it('should call handletoggle with argument "startRec" when clicked and recStatus is "off" and tab is valid', () => {
-    wrapper = shallow(<ToggleButton recStatus='off' isValidTab={true} handleToggle={handleToggleMock}/>)
+    wrapper = shallow(<ToggleButton recStatus="off" isValidTab handleToggle={handleToggleMock} />);
     expect(wrapper).toMatchSnapshot();
     const button = wrapper.find('button');
     button.simulate('click');
     expect(handleToggleMock).toHaveBeenCalledWith(ControlAction.START);
   });
   it('should call handletoggle with argument "stopRec" when clicked and recStatus is "on"', () => {
-    wrapper = shallow(<ToggleButton recStatus='on' isValidTab={true} handleToggle={handleToggleMock}/>)
+    wrapper = shallow(<ToggleButton recStatus="on" isValidTab handleToggle={handleToggleMock} />);
     expect(wrapper).toMatchSnapshot();
     const button = wrapper.find('button');
     button.simulate('click');
     expect(handleToggleMock).toHaveBeenCalledWith(ControlAction.STOP);
   });
   it('should call handletoggle with argument "startRec" when clicked and recStatus is "paused"', () => {
-    wrapper = shallow(<ToggleButton recStatus='paused' isValidTab={true} handleToggle={handleToggleMock}/>)
+    wrapper = shallow(<ToggleButton recStatus="paused" isValidTab handleToggle={handleToggleMock} />);
     expect(wrapper).toMatchSnapshot();
     const button = wrapper.find('button');
     button.simulate('click');
     expect(handleToggleMock).toHaveBeenCalledWith(ControlAction.START);
   });
-})
+});
